pyproject.toml
[project]
name = "tribia-python-assessment"
version = "0.1.0"
description = ""
authors = [
    {name = "Pituccio",email = "pedrolino.reynolds@proton.me"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "playwright (>=1.50.0,<2.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
